implement serialization for context-related types
