+++ stack-core/src/vec_one.rs
@@ -3,11 +3,20 @@
 use core::slice::{Iter, IterMut, SliceIndex};
 use std::vec::IntoIter;
 
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 /// A [`Vec`] with at least one element.
 #[derive(
-  Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Deserialize,
+  Debug,
+  Clone,
+  PartialEq,
+  Eq,
+  PartialOrd,
+  Ord,
+  Hash,
+  Default,
+  Serialize,
+  Deserialize,
 )]
 pub(crate) struct VecOne<T> {
   vec: Vec<T>,
